Reorganize organization mutations
