--- app/graphql/mutations/delete_organization_mutation.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  # GraphQL mutation to delete an organization
-  class DeleteOrganizationMutation
-    def call(organization, _arguments, _context)
-      organization.destroy
-      true
-    end
-  end
-end
