--- app/graphql/mutations/save_organization_mutation.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  # GraphQL mutation to update an organization
-  class SaveOrganizationMutation
-    def call(organization, arguments, _context)
-      params = arguments[:data].to_h.compact
-      params['description'] = nil if params['description'].empty?
-      organization.update(params)
-      organization
-    end
-  end
-end
